Move dialogs to subpackage
